+++ changelog.md
@@ -2,3 +2,4 @@
 * Started the plugin structure, added a whole bunch of pre-existing MailChimp code
 * Added composer requirements
 * Generate a webhook url + key
+* Added checkbox to user profile to confirm MailChimp syncing